create entity Team adn DTO team

+++ src/main/java/it/contrader/dto/TeamDTO.java
@@ -20,7 +20,7 @@ public class TeamDTO {
 
 	private int numPlayers;
 
-	private float rating;
+	private int rating;
 
 	private int homeWins;
